+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.11.0] - 2023-10-25
 ### Added
 - `Helper` Added method to get the hit target using a raycast to the CameraHelper
+- `Types` Added attribute to validate that a gameobject has a specific component
+
+### Dependencies
+- Added support for Odin Validator (not required)
 
 ## [0.10.0] - 2023-10-14
 ### Added